[ng] vars: separate config headers in two vars (local and non-local ones)
Where the "local" are those which are in the same subdirectory of the
Makefile.am being currently processed, and the "non-local" ones are
not (and thus need to be prepended with $(top_builddir)).

* automake.in (handle_config_headers): Define two new make variables
'am.config-hdr.local' and 'am.config-hdr.non-local'.  Accordingly,
define 'AM_CONFIG_HEADERS' as the union of them.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>