[ng] general: new internal vars to hold path of current Makefile{,.in,.am}
This is preferable to having to use the '%MAKEFILE%', '%MAKEFILE-IN%'
and '%MAKEFILE-AM%' transforms in several places, and will enable us
to do more sweeping refactorings in the future.

* automake.in (generate_makefile): Define new private make variables:
  - am.relpath.makefile.am
  - am.relpath.makefile.in
  - am.relpath.makefile
* automake.in (handle_configure, handle_clean, generate_makefile),
lib/am/configure.am, lib/am/check-typos.am, lib/am/clean.am: Adjust
throughout to use them instead of the corresponding transforms.
* t/remake.sh: Adjust and extend grepping checks.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>