[ng] tests: remove dead code from a test, fix a typo
* t/dist-formats.tap: Fix typo in a message: s/.tar.Z/.tar.xz/.
(have_compressor): Remove special handling of the 'compress' tool.
This is now useless, since we've removed support for '.tar.Z' tarballs
and thus care no more whether a 'compress' program is available or not.
Also, in the main code, one stray remaining reference to the 'shar'
program, that too useless now that we've removed support for '.shar.gz'
archives.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>