[ng] dist: new API to specify formats of distribution tarballs
The API to specify the formats of distribution tarballs has been changed
completely.

Instead of using the various 'dist-*' automake options, the developer is
now expected to specify the default formats of its distribution tarballs
with the special variable AM_DIST_FORMATS; for example, where once would
have been:

    AM_INIT_AUTOMAKE([dist-bzip2 dist-zip])

now it must be:

    AM_DIST_FORMATS = gzip bzip2 zip

and similarly, where once would have been:

    AUTOMAKE_OPTIONS = no-dist-gzip dist-bzip2 dist-xz

now is it simply:

    AM_DIST_FORMATS = bzip2 xz

Similarly, The various 'dist-*' targets (dist-gzip, dist-xz, dist-lzip,
dist-bzip2 and dist-gz) has been removed.  If the user wants to generate
tarballs for formats not specified by the default AM_DIST_FORMATS, he can
simply override that variable dynamically:

    # Will generate a '.zip' archive and a '.tar.xz' archive, and not
    # further ones.
    make dist AM_DIST_FORMATS='zip xz'

This change is of course is totally backward incompatible, but the
enhanced flexibility and simplicity is worth it.  Not to mention that
the transition from the mainline Automake API to the new Automake-NG
one is trivial.

* NG-NEWS: Update.
* ng/automake-ng.texi: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>