readme: fixlets to HACKING
* HACKING: Use longer "=====" lines to separate different section (this
is just eye-candy admittedly, but I prefer it).
(Release procedure): Don't tell to "update NEWS"; that should be updated
throughout the normal course of development.  Instead, tell to just check
it.  Improve description of the re-bootstrapping and rechecking procedure,
also suggesting to use "git clean" beforehand (with all due warnings!).

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>