compile: remove support for $(INCLUDES)
It has already been deprecated in the manual and by warnings in the
'obsolete' category for ages (at least since 2003), in favour of
AM_CPPFLAGS.  Automake-NG has removed support for it already.  And
the warnings in the obsolete category are on by default since the
1.12.4 release.

So, by removing it in Automake 1.14, we will simplify the transition
path for people that want to switch to Automake-NG.

See also Automake-NG commit v1.12-267-g6b74968 of 2012-05-21,
"[ng] compile: don't support $(INCLUDES) anymore, it's obsolete".

* NEWS, doc/automake.texi: Update.
* automake.in (generate_makefile): Reject INCLUDES unconditionally.
In all languages: don't add $(INCLUDES) to the compiler command line.
* t/includes-deprecation.sh: New test.
* t/list-of-tests.mk: Add it.
* t/werror2.sh : Adjust to avoid spurious failures.
* t/pluseq5.sh: Likewise.
* t/warnopts.sh: Likewise.

Suggested-by: Paolo Bonzini <bonzini@gnu.org>
Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>