coverage: expose fork bomb in $(BUILT_SOURCES) handling
Due to how the handling of $(BUILT_SOURCES) is implemented in Automake-NG,
a recursive make call in the recipe of any $(BUILT_SOURCES) (or of any of
its prerequisites) will cause an infinite recursion (complete with fork
bomb, yuck).  Expose the issue in a test case (still failing, of course).

See:
<http://lists.gnu.org/archive/html/help-smalltalk/2012-08/msg00027.html>
<http://lists.gnu.org/archive/html/automake-patches/2012-08/msg00052.html>

* t/built-sources-fork-bomb.sh: New test.
* Makefile.am (XFAIL_TESTS): Add it.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>