depcomp: remove 'icc' depmode, add 'tcc' one
According to the comments in the now-removed test 'depcomp5.test' (removed
by commit 'v1.11-1829-g5066c1b' of 2012-02-08, "tests: improve and rework
tests on dependency tracking") "icc 8.0 and greater understand gcc options,
so depmode is set to gcc".  According to Wikipedia, the 8.0 version of the
Intel C compiler has been released in December 2003, that is, almost nine
yeas ago.  So we can remove explicit support for icc 7.x from the depcomp
script.

Since our support for tcc has so far been relied on the icc depmode, we
must write a new explicit 'tcc' depmode.  That is easy to do, since we can
rely on the simpler and better-specified format of tcc-specified makefiles;
for more info, see:
<http://lists.gnu.org/archive/html/automake/2012-10/msg00043.html>

* lib/depcomp (icc): Remove.
(tcc): New.
* NEWS: Update.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>