tests: ensure generation of wrapper tests matching multiple conditions
* gen-testsuite-part: Our old code to generate wrapper tests had a
severe limitation, in that if a test matched two or more conditions
calling for generation of wrapper tests, still only one wrapper test
was generated, instead of the three that would have been expected --
that is, one using the setup code triggered by the first condition,
one using the setup code triggered by the second condition, and one
using both this setup code fragments.
Admittedly, this was only a theoretical limitation for the moment,
since since so far no test exists that matches two or more conditions
for wrapping.  Still, this might change in the future, and easily in
an unnoticed way, so better fix the issue now, before it might become
a real problem.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>