vala: if no proper compiler found, set $(VALAC) to 'valac'
This is better than setting it to ':' (as is currently done), because
a triggered makefile rule invoking a vala compilation will then clearly
fail with an informative error message like "valac: command not found",
rather than silently, with the error possibly going unnoticed, or
triggering harder-to-diagnose fallout failures in later steps.

For a precedent of a similar behaviour, see the AC_PROG_YACC macro.

* m4/valac.m4: Implement the new semantic.
* doc/automake.texi (Vala Support): Document it.
* t/vala4.sh: Adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>