tests: prefer including 'test-init.sh' rather than './defs'
This is a follow-up to today's same-name commit v1.12.4-181-g5ddf100,
which did the same for the test cases in master.

* All tests: To run the common setup, use the command:

    . test-init.sh

instead of the older, "historical" one:

    . ./defs || exit 1

Note that the "|| exit 1" wasn't really useful, since the 'errexit'
shell flag is in effect in both './defs' and 'test-init.sh', and all
the known shells that are good enough to run the automake testsuite
do automatically exit with error when a sourced file cannot be found
(at least, they do so in non-interactive mode, which is the only
mode that concerns us in the testsuite).

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>