vala tests: don't use the 'posix' profile, it's no longer supported
Fixes automake bug#12934 a.k.a. bug#12522.

Some of automake's Vala tests used to rely on "valac --profile posix"
(which makes it not use glib).  However, the posix profile was removed
in August 2012, and is no longer present in the GNOME 3.6.2 version
of valac (see commit ca020bf0 in the vala Git repository).

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>

* t/vala-libs.sh: Adjust to avoid using the posix profile.
* t/vala-mix.sh: Likewise.
* t/vala-mix2.sh: Likewise.
* t/vala-parallel.sh: Likewise.
* t/vala-vapi.sh: Likewise.
* t/vala-vpath.sh: Likewise.
* NEWS: Update and adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>