tests: better names for temporary configure.ac files
When editing a pre-existent 'configure.ac' file in a test script, name
the temporary file 'configure.tmp' rather than 'configure.int'; the
latter was a relict of the times when we used 'configure.in' as the
default name for the Autoconf input files throughout the testsuite.

* t/cond43.sh: Adjust.
* t/depcomp8a.sh: Likewise.
* t/depcomp8b.sh: Likewise.
* t/gettext.sh: Likewise.
* t/mmode.sh: Likewise.
* t/pr401.sh: Likewise.
* t/pr401b.sh: Likewise.
* t/pr401c.sh: Likewise.
* t/python11.sh: Likewise.
* t/cond42.sh: Likewise.  Also, while at it, rename ...
(edit_configure_in): ... this function ...
(edit_configure_ac): ... like this.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>