obsolete: better error message if AM_PROG_CC_STDC is used
In Automake 1.13, the long-deprecated (since 2002) macro AM_PROG_CC_STDC
has been removed.  Such a removal, albeit sensible and justified, was
probably done in a too-abrupt way, since it didn't turn the pre-existing
warning messages into fatal error messages, but simply dropped the macro
definition, so that remaining usages of it would cause unclear error
messages, e.g.:

    configure.ac:4: warning: macro 'AM_PROG_CC_STDC' not found in library

from aclocal, and:

    configure.ac:4: error: possibly undefined macro: AM_PROG_CC_STDC
      If this token and others are legitimate, please use m4_pattern_allow.
      See the Autoconf documentation.

from autoconf.

In an attempt to mitigate this issue, we re-add an AM_PROG_CC_STDC
definition that simply raises a *clear* error message when the macro
is used.

* t/am-prog-cc-stdc-no-more.sh: New test.
* t/list-of-tests.mk: Add it.
* m4/obsolete-err.m4: Add the new "error-raising" definition for
AM_PROG_CC_STDC; the error message is a variation of the one already
present in the older version of this macro, before it got removed in
commit 'v1.12-15-gd2ca168'.
* NEWS: Update.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>