tests: don't always look for a C++ compiler named 'RCC'
On MacOS X (10.8), since the file system is case-insensitive, RCC
can point to the "Resource Compiler" of the Qt4 Toolkit:

    <http://doc.qt.digia.com/4.2/rcc.html>

That mismatch causes our configure script to erroneously think that
no working C++ compiler is present, and that is thus necessary to
skip all the test cases requiring such a compiler.

So only look for a compiler named 'RCC' if the file system is
case-sensible.

Issue spotted analyzing the testsuite logs reported in bug#13317.

* configure.ac: Adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>